--- client/src/components/LoginBtn.js
@@ -1,14 +1,0 @@
-import React from "react";
-import Button from "@mui/material/Button";
-
-function LoginBtn() {
-  return (
-    <div>
-      <Button href="/login" variant="contained">
-        Sign Up/Login
-      </Button>
-    </div>
-  );
-}
-
-export default LoginBtn;
