+++ client/src/components/LoginButton.js
@@ -1,0 +1,11 @@
+import React from "react";
+import MaterialButton from "./MaterialButton";
+
+export default function LoginButton() {
+  return (
+    <div>
+      <MaterialButton href="/login" text="Sign Up/Login"></MaterialButton>
+    </div>
+  );
+}
+
