+++ client/src/components/LogoutButton.js
@@ -1,0 +1,11 @@
+import React from "react";
+import MaterialButton from "./MaterialButton";
+
+export default function LogoutButton() {
+    return (
+      <div>
+        <MaterialButton href= "/logout" text = "Logout"></MaterialButton>
+      </div>
+    );
+  }
+
